Changed sceneRadius from float to double to increase calculation precision.

+++ include/openspace/scene/scenegraphnode.h
@@ -94,8 +94,8 @@ public:
     const Renderable* renderable() const;
     Renderable* renderable();
 
-    void setSceneRadius(float SceneRadius);
-    const float& sceneRadius() const;
+    void setSceneRadius(double SceneRadius);
+    const double& sceneRadius() const;
 
     // @TODO Remove once the scalegraph is in effect ---abock
     void setEphemeris(Ephemeris* eph) {
@@ -109,7 +109,7 @@ private:
     std::vector<SceneGraphNode*> _children;
     SceneGraphNode* _parent;
     Ephemeris* _ephemeris;
-    float _sceneRadius;
+    double _sceneRadius;
 
     PerformanceRecord _performanceRecord;
